Wrote acceptance tests for BankScroll and stubs for BankScroll class
The BankScrollTest class checks that when used, a BankScroll returns
a reminder to the player about the contents of their bank chest, and
it also checks that the BankScroll remains in the inventory of the
player after use. To be able to run the test I created stubs for
the BankScroll constructor and also for the useScroll method.