Implemented Lon Jatham NPC and the Java Quiz
All tests now pass. Lon Jatham now exists in the required place in the
game and has a sprite and gives the Java Quiz (which now exists as a
quest) when prompted. Given a correct answer he gives XP and a
certificate . Certificate now exists as a document Item and as a word in
the game.