Updated bank tellers tests
For each of the three bank tellers there are now tests that check the
mark command. When the player is equipped with some empty scrolls, the
tests check that after the mark command is done, the number of empty
scrolls is decremented by 1 and the number of bank scrolls is
incremented by 1. The test also checks that the new bank scroll is
marked with the correct bank. When the player doesn't have any empty
scrolls the tests check the reply from the teller.