Implemented the mark command for Zara and the bank tellers
The bank tellers and Zara now have the mark command that can be used by
the player to create a bank scroll for the corresponding banks (and
Zara's chest) that will display the items inside the chests. The command
is successful only if the user is equipped with at least one empty
scroll. A bank scroll for Zara's chest is created only if the player has
completed the Suntan Cream for Zara quest.