Updated tests for the logbook check
Now the tests check what happens if the player has/hasn't got a logbook
equipped when they speak to Lon Jatham. They make sure that Lon does not
speak back to the player if the user does not have their logbook in
their inventory. The tests also make sure that Granny Graham in the
cottage gives the user one logbook. 
In code review, some minor typo changes i.e. test -> #test, to indicate
to the user what to say in response to Lon.