Wrote tests for ItemsOnTable, ItemsUpstairs and ToysCollector
Added two new test classes for the items downstairs and upstairs in the
Ross house. ItemsOnTableTest checks if there is a portal to upstairs
whilst ItemsUpstairsTest checks if there is a portal to downstairs and
if there is a chest containing a teddy. I also added a test to the
ToysCollector class which tests the trigger phrase for the teddy.

Issue: #10