Implemented logbook with Lon Jatham
Now the user can only speak to Lon Jatham if they are holding their
logbook. A logbook can easily be acquired by talking to Granny Graham
(HouseKeeperNPC). The Java quiz can only be started if the user has
their logbook, and if the user drops their logbook at any time they try
to speak to Lon he will not engage in conversation. I have updated the
book.xml to include the details about the logbook. Also fixed a typo in
words.txt which fixes the plural spelling of 'certificates'.