Changes to code in response to code review
Have separated out the test cases, as this has been highlighted as a
better code practice in code review, and therefore added comments before
each test case to describe what each method is testing.
Added a string for the "logbook" so this can be changed if the name of
the book changes in future revisions of the game code.
Removed the unnecessary imports in LonJathamNPC.java and JavaQuiz.java.
Fixed the HouseKeeperNPCTest.java bug so it nows checks correctly that
the user only has 1 logbook equipped.