[#18] Updated DefaultAction, ActionXMLLoader and SlashActionRepository
Updated DefaultAction to implement the SlashAction interface and
implemented the 'execute' method that sets the type, parameters and
remainder of the action and sends it to the server. Also updated the
constructor for this class and created getters and setters for the type,
parameters, modifier, minParameters and maxParameters.

Modified ActionXMLLoader to match the new version of DefaultAction.

Modified SlashActionRepository to load all existing xml actions and add
them in a HashTable as DefaultActions. The three existing xml actions
(Teleport, Invisible and Message) are also updated in the HashTable with
the corresponding parameters and remainder.