Moved loaders from server to client
ActionXMLLoader and ActionGroupsXMLLoader were initially in a server
package but should have been in a client package. I have now fixed this.
However, GroupsXMLLoader is private and cannot be accessed in this
package so as a temporary fix I have made a copy of this. O have a;sp
updated the packet names in SlashActionRepository.