[#19] Change use of remainder for actions to check for empty remainder
Changed the xml schema for actions so that if an action has a remainder
then it also needs to specify if that remainder can be empty or not (for
actions such as grumpy). Added this to all actions which have remainders
and changed the ActionXMLLoader and DefaultAction class in order to make
them use this value when setting the remainder of actions.